requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
build==0.10.0
CacheControl==0.12.14
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
crashtest==0.4.1
cryptography==38.0.4
dataclasses-json==0.5.7
debugpy==1.6.5
discord.py==2.1.0
distlib==0.3.7
dulwich==0.21.5
faiss-cpu==1.7.3
filelock==3.9.0
Flask==2.2.2
Flask-Inputs==0.3.0
frozenlist==1.3.3
future==0.18.3
geojson==2.5.0
googletrans==4.0.0rc1
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.8.0
installer==0.7.0
iso8601==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
keyring==23.13.1
langchain==0.0.125
langdetect==1.0.9
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.1
openai==0.27.2
packaging==23.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.0.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.21.12
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.16.0
pydantic==1.9.2
pyflakes==2.5.0
pyowm==3.3.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyseto==1.7.0
PySocks==1.7.1
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
pytoolconfig==1.2.5
PyYAML==6.0
rapidfuzz==2.15.1
referencing==0.30.2
regex==2022.10.31
replit==3.2.5
replit-python-lsp-server==1.15.9
requests==2.28.2
requests-toolbelt==1.0.0
rfc3986==1.5.0
rope==1.7.0
rpds-py==0.9.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
tenacity==8.2.0
tiktoken==0.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.64.1
trove-classifiers==2023.8.7
typing-extensions==3.10.0.2
typing-inspect==0.8.0
ujson==5.7.0
urllib3==1.26.14
virtualenv==20.24.3
webencodings==0.5.1
Werkzeug==2.2.2
whatthepatch==1.0.3
WTForms==3.0.1
xattr==0.10.1
yapf==0.32.0
yarl==1.8.2
zipp==3.16.2
